Add recipients to write step, persist selections in draft.
