+++ components/publish/ShareTestimony.tsx
@@ -2,87 +2,33 @@ import { faCopy } from "@fortawesome/free-solid-svg-icons"
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome"
 import clsx from "clsx"
 import { External, maple } from "components/links"
-import { isNotNull } from "components/utils"
-import { cloneDeep, fromPairs, isString, last, sortBy } from "lodash"
 import { useRouter } from "next/router"
-import { useCallback, useEffect, useState } from "react"
-import { components, GroupBase, MultiValueGenericProps } from "react-select"
+import { useCallback, useState } from "react"
 import styled from "styled-components"
 import { Button, Modal } from "../bootstrap"
 import { CopyButton } from "../buttons"
-import { useMemberSearch } from "../db"
-import { useProfileState } from "../db/profile/redux"
 import { useAppDispatch } from "../hooks"
-import { Loading, MultiSearch } from "../legislatorSearch"
-import { calloutLabels } from "./content"
 import { useFormRedirection, usePublishState, useTestimonyEmail } from "./hooks"
 import * as nav from "./NavigationButtons"
-import {
-  addCommittee,
-  addMyLegislators,
-  clearLegislatorSearch,
-  Legislator,
-  resolvedLegislatorSearch,
-  setRecipients,
-  setShowThankYou
-} from "./redux"
+import { setShowThankYou } from "./redux"
+import { SelectRecipients } from "./SelectRecipients"
 import { SendEmailButton } from "./SendEmailButton"
 import { StepHeader } from "./StepHeader"
 import { YourTestimony } from "./TestimonyPreview"
 
 export const ShareTestimony = styled(({ ...rest }) => {
   useFormRedirection()
-  useEmailRecipients()
   return (
     <div {...rest}>
-      <StepHeader step={3}>Send Your Email</StepHeader>
-      <p className="fw-bold mt-4">Let Your Voice Be Heard!</p>
-      <YourTestimony className="position-relative">
+      <StepHeader step={3}>Confirm and Send</StepHeader>
+      <SelectRecipients className="mt-4" />
+      <YourTestimony className="mt-4 position-relative">
         <CopyTestimony />
       </YourTestimony>
-      <RecipientControls className="mt-5" />
-      <SelectLegislatorEmails className="mt-2" />
       <nav.FormNavigation right={<ShareButtons />} />
     </div>
   )
-})`
-  .label-callout {
-    font-size: 0.75rem;
-    font-style: italic;
-    text-align: center;
-  }
-
-  .leg-search__control {
-    background-color: var(--bs-body-bg);
-    border-color: var(--bs-body-bg);
-    background-image: url("/mail-bg.svg");
-    background-repeat: no-repeat;
-    background-position: bottom right 20%;
-    background-size: auto 6rem;
-    min-height: 10rem;
-    border-radius: 1rem;
-  }
-
-  .leg-search__value-container {
-    padding: 0.5rem;
-    align-self: flex-start;
-  }
-
-  .leg-search__multi-value {
-    background-color: var(--bs-blue);
-  }
-
-  .leg-search__multi-value__label,
-  .leg-search__multi-value__remove {
-    color: white;
-  }
-
-  .leg-search__multi-value,
-  .leg-search__multi-value__remove,
-  .leg-search__control {
-    border-radius: 1rem;
-  }
-`
+})``
 
 const CopyTestimony = styled(props => {
   const email = useTestimonyEmail()
@@ -114,147 +60,6 @@ const EmailHelp = (props: { className?: string }) => {
       Having trouble opening the email draft?
     </External>
   )
-}
-
-const RecipientControls = styled(({ className }) => {
-  const dispatch = useAppDispatch()
-  const { share } = usePublishState()
-
-  const email = useTestimonyEmail()
-  const buttons = []
-
-  if (email.to)
-    buttons.push(
-      <CopyButton key="copy" variant="link" text={email.to}>
-        <FontAwesomeIcon icon={faCopy} /> Copy Recipients
-      </CopyButton>
-    )
-
-  if (share.committeeChairs.length > 0)
-    buttons.push(
-      <Button
-        key="committee"
-        variant="link"
-        onClick={() => dispatch(addCommittee())}
-      >
-        Add Relevant Committee
-      </Button>
-    )
-
-  if (share.userLegislators.length > 0)
-    buttons.push(
-      <Button
-        key="legislators"
-        variant="link"
-        onClick={() => dispatch(addMyLegislators())}
-      >
-        Add My Legislators
-      </Button>
-    )
-
-  return (
-    <div className={clsx("d-flex gap-4 justify-content-end", className)}>
-      {buttons}
-    </div>
-  )
-})`
-  button {
-    font-size: 0.75rem;
-    padding: 0;
-    justify-self: flex-end;
-  }
-`
-
-const Label = (
-  props: MultiValueGenericProps<Legislator, true, GroupBase<Legislator>>
-) => {
-  const label = props.children
-  const callout = props.data.callout
-  return (
-    <components.MultiValueLabel {...props}>
-      {label}
-      <div className="label-callout">{callout}</div>
-    </components.MultiValueLabel>
-  )
-}
-
-const SelectLegislatorEmails = ({ className }: { className?: string }) => {
-  const { share } = usePublishState()
-  const dispatch = useAppDispatch()
-  return share.loading ? (
-    <Loading />
-  ) : (
-    <MultiSearch
-      className={className}
-      placeholder="Find your legislators"
-      index={share.options}
-      memberIds={share.recipients.map(m => m.MemberCode)}
-      update={m => dispatch(setRecipients(m))}
-      getOptionLabel={o =>
-        `${o.Branch === "House" ? "Rep." : "Sen."} ${o.Name} | ${o.District}`
-      }
-      components={{ MultiValueLabel: Label }}
-    />
-  )
-}
-
-const useEmailRecipients = () => {
-  const { bill: { currentCommittee } = {} } = usePublishState()
-  const { profile: { representative, senator } = {} } = useProfileState()
-  const { index } = useMemberSearch()
-  const dispatch = useAppDispatch()
-
-  useEffect(() => {
-    dispatch(clearLegislatorSearch())
-  }, [dispatch])
-
-  useEffect(() => {
-    if (!index) return
-
-    const legislatorsById: Record<string, Legislator> = fromPairs(
-      cloneDeep([...index.representatives, ...index.senators]).map(m => [
-        m.MemberCode,
-        m
-      ])
-    )
-
-    const committeeCallouts = [
-        [currentCommittee?.houseChair?.id, calloutLabels.houseChair],
-        [currentCommittee?.senateChair?.id, calloutLabels.senateChair]
-      ],
-      userLegislatorsCallouts = [
-        [representative?.id, calloutLabels.yourRepresentative],
-        [senator?.id, calloutLabels.yourSenator]
-      ],
-      callouts = [...committeeCallouts, ...userLegislatorsCallouts]
-
-    callouts.forEach(([id, callout]) => {
-      if (!id) return
-      const member = legislatorsById[id]
-      if (member) {
-        if (member.callout) member.callout += ", " + callout
-        else member.callout = callout
-      }
-    })
-
-    dispatch(
-      resolvedLegislatorSearch({
-        options: sortBy(Object.values(legislatorsById), m =>
-          last(m.Name.split(" "))
-        ),
-        committeeChairs: committeeCallouts
-          .map(([id]) => id)
-          .filter(isString)
-          .map(id => legislatorsById[id])
-          .filter(isNotNull),
-        userLegislators: userLegislatorsCallouts
-          .map(([id]) => id)
-          .filter(isString)
-          .map(id => legislatorsById[id])
-          .filter(isNotNull)
-      })
-    )
-  }, [currentCommittee, dispatch, index, representative, senator])
 }
 
 const ShareButtons = () => {
