+++ components/publish/hooks/useFormSync.ts
@@ -82,12 +82,18 @@ function useSyncTestimonyToStore(edit: UseEditTestimony) {
 
 type DraftContent = ReturnType<typeof useFormDraft>
 function useFormDraft() {
-  const { attachmentId, content, position } = usePublishState()
-  return { attachmentId, content, position }
+  const { attachmentId, content, position, recipientMemberCodes } =
+    usePublishState()
+  return { attachmentId, content, position, recipientMemberCodes }
 }
 
 function usePersistedDraft(draft?: WorkingDraft): DraftContent | undefined {
   if (!draft) return
-  const { attachmentId, content, position } = draft
-  return { attachmentId: attachmentId ?? undefined, content, position }
+  const { attachmentId, content, position, recipientMemberCodes } = draft
+  return {
+    attachmentId: attachmentId ?? undefined,
+    content,
+    position,
+    recipientMemberCodes: recipientMemberCodes ?? undefined
+  }
 }
